Fix NaN score
